==> 1_data_preprocessing.R <==

####### source data #######
#setwd("/home/data/t210447/r_workplace/driver_gene/")
#####[1]load data
#CRC_GSE146771_data <- readRDS("CRC_GSE146771_Smartseq2_expression.rds")
#CRC_EMTAB8107_data <- readRDS("CRC_EMTAB8107_expression.rds")
CRC_GSE166555_data <- readRDS("CRC_GSE166555_expression.rds")

#####[2]Load filtering data of the malignant cell population
#CRC_GSE146771
CRC_GSE146771_MC <- read.csv("CRC_GSE146771_Smartseq2_Malignant_cell_14_17_9.csv", header = TRUE, sep = ",")
CRC_GSE146771_MC_C14 <- CRC_GSE146771_MC[1:269,1]#C14 malignant cell population name
CRC_GSE146771_MC_C17 <- CRC_GSE146771_MC[270:479,1]#C17 malignant cell population name
CRC_GSE146771_MC_C9 <- CRC_GSE146771_MC[480:1057,1]#C9 malignant cell population name
CRC_GSE146771_data_C14 <- CRC_GSE146771_data[,CRC_GSE146771_MC_C14]#C14 malignant cell population data
CRC_GSE146771_data_C17 <- CRC_GSE146771_data[,CRC_GSE146771_MC_C17]#C17 malignant cell population data
CRC_GSE146771_data_C9 <- CRC_GSE146771_data[,CRC_GSE146771_MC_C9]#C9 malignant cell population data

#CRC_EMTAB8107
CRC_EMTAB8107_MC <- read.csv("CRC_EMTAB8107_Malignant_cell_6_8_9.csv", header = TRUE, sep = ",")
CRC_EMTAB8107_MC_C6 <- CRC_EMTAB8107_MC[1:1961,1]#C6 malignant cell population name
CRC_EMTAB8107_MC_C8 <- CRC_EMTAB8107_MC[1962:3747,1]# C8 malignant cell population name
CRC_EMTAB8107_MC_C9 <- CRC_EMTAB8107_MC[3748:5497,1]#C9 malignant cell population name
CRC_EMTAB8107_data_C6 <- CRC_EMTAB8107_data[,CRC_EMTAB8107_MC_C6]#C6 malignant cell population data
CRC_EMTAB8107_data_C8 <- CRC_EMTAB8107_data[,CRC_EMTAB8107_MC_C8]#C8 malignant cell population data
CRC_EMTAB8107_data_C9 <- CRC_EMTAB8107_data[,CRC_EMTAB8107_MC_C9]#C9 malignant cell population data

#CRC_GSE166555
CRC_GSE166555_MC <- read.csv("CRC_GSE166555_Malignant_cell_5_11_13_14_18_29.csv", header = TRUE, sep = ",")
CRC_GSE166555_MC_C5 <- CRC_GSE166555_MC[1:1484,1]#C5 malignant cell population name
CRC_GSE166555_MC_C11 <- CRC_GSE166555_MC[1485:3881,1]#C11 malignant cell population name
CRC_GSE166555_MC_C13 <- CRC_GSE166555_MC[3882:5545,1]#C13 malignant cell population name
CRC_GSE166555_MC_C14 <- CRC_GSE166555_MC[5546:7500,1]#C14 malignant cell population name
CRC_GSE166555_MC_C18 <- CRC_GSE166555_MC[7501:7839,1]#C18 malignant cell population name
CRC_GSE166555_MC_C29 <- CRC_GSE166555_MC[7840:8135,1]#C29 malignant cell population name
CRC_GSE166555_data_C5 <- CRC_GSE166555_data[,CRC_GSE166555_MC_C5]#C5 malignant cell population data
CRC_GSE166555_data_C11 <- CRC_GSE166555_data[,CRC_GSE166555_MC_C11]#C11 malignant cell population data
CRC_GSE166555_data_C13 <- CRC_GSE166555_data[,CRC_GSE166555_MC_C13]#C13 malignant cell population data
CRC_GSE166555_data_C14 <- CRC_GSE166555_data[,CRC_GSE166555_MC_C14]#C14 malignant cell population data
CRC_GSE166555_data_C18 <- CRC_GSE166555_data[,CRC_GSE166555_MC_C18]#C18 malignant cell population data
CRC_GSE166555_data_C29 <- CRC_GSE166555_data[,CRC_GSE166555_MC_C29]#C29 malignant cell population data

#####[3]Load differentially expressed gene (DEG) data of the filtered malignant cell population
#CRC_GSE146771
CRC_GSE146771_DE <- read.csv("CRC_GSE146771_Smartseq2_DE_cluster_14_17_9.csv", header = TRUE, sep = ",")
CRC_GSE146771_DE_C14 <- CRC_GSE146771_DE[1:1887,2]#C14 DEGs name
CRC_GSE146771_DE_C17 <- CRC_GSE146771_DE[1888:4439,2]#C17 DEGs name
CRC_GSE146771_DE_C9 <- CRC_GSE146771_DE[4440:6974,2]#C9 DEGs name
CRC_GSE146771_data_DE_C14 <- CRC_GSE146771_data[CRC_GSE146771_DE_C14,CRC_GSE146771_MC_C14]#C14 DEGs data
CRC_GSE146771_data_DE_C17 <- CRC_GSE146771_data[CRC_GSE146771_DE_C17,CRC_GSE146771_MC_C17]#C17 DEGs data
CRC_GSE146771_data_DE_C9 <- CRC_GSE146771_data[CRC_GSE146771_DE_C9,CRC_GSE146771_MC_C9]#C9 DEGs data

#CRC_EMTAB8107
CRC_EMTAB8107_DE <- read.csv("CRC_EMTAB8107_DE_cluster_6_8_9.csv", header = TRUE, sep = ",")
CRC_EMTAB8107_DE_C6 <- CRC_EMTAB8107_DE[1:1670,2]#C6 DEGs name
CRC_EMTAB8107_DE_C8 <- CRC_EMTAB8107_DE[1671:2984,2]#C8 DEGs name
CRC_EMTAB8107_DE_C9 <- CRC_EMTAB8107_DE[2985:4483,2]#C9 DEGs name
CRC_EMTAB8107_data_DE_C6 <- CRC_EMTAB8107_data[CRC_EMTAB8107_DE_C6,CRC_EMTAB8107_MC_C6]#C6 DEGs data
CRC_EMTAB8107_data_DE_C8 <- CRC_EMTAB8107_data[CRC_EMTAB8107_DE_C8,CRC_EMTAB8107_MC_C8]#C8 DEGs data
CRC_EMTAB8107_data_DE_C9 <- CRC_EMTAB8107_data[CRC_EMTAB8107_DE_C9,CRC_EMTAB8107_MC_C9]#C9 DEGs data

#CRC_GSE166555
CRC_GSE166555_DE <- read.csv("CRC_GSE166555_DE_cluster_5_11_13_14_18_29.csv", header = TRUE, sep = ",")
CRC_GSE166555_DE_C5 <- CRC_GSE166555_DE[7486:8832,2]#C5 DEGs name
CRC_GSE166555_DE_C11 <- CRC_GSE166555_DE[1:1345,2]#C11 DEGs name
CRC_GSE166555_DE_C13 <- CRC_GSE166555_DE[1346:2677,2]#C13 DEGs name
CRC_GSE166555_DE_C14 <- CRC_GSE166555_DE[2678:4881,2]#C14 DEGs name
CRC_GSE166555_DE_C18 <- CRC_GSE166555_DE[4882:6516,2]#C18 DEGs name
CRC_GSE166555_DE_C29 <- CRC_GSE166555_DE[6517:7485,2]#C29 DEGs name
CRC_GSE166555_data_DE_C5 <- CRC_GSE166555_data[CRC_GSE166555_DE_C5,CRC_GSE166555_MC_C5]#C5 DEGs data
CRC_GSE166555_data_DE_C11 <- CRC_GSE166555_data[CRC_GSE166555_DE_C11,CRC_GSE166555_MC_C11]#C11 DEGs data
CRC_GSE166555_data_DE_C13 <- CRC_GSE166555_data[CRC_GSE166555_DE_C13,CRC_GSE166555_MC_C13]#C13 DEGs data
CRC_GSE166555_data_DE_C14 <- CRC_GSE166555_data[CRC_GSE166555_DE_C14,CRC_GSE166555_MC_C14]#C14 DEGs data
CRC_GSE166555_data_DE_C18 <- CRC_GSE166555_data[CRC_GSE166555_DE_C18,CRC_GSE166555_MC_C18]#C18 DEGs data
CRC_GSE166555_data_DE_C29 <- CRC_GSE166555_data[CRC_GSE166555_DE_C29,CRC_GSE166555_MC_C29]#C29 DEGs data

#save - load RData
save(CRC_GSE146771_data_DE_C14,CRC_GSE146771_data_DE_C17,CRC_GSE146771_data_DE_C9, file = "CRC_GSE146771_data_DE_C14_17_9.RData")
load("CRC_GSE146771_data_DE_C14_17_9.RData")
save(CRC_EMTAB8107_data_DE_C6,CRC_EMTAB8107_data_DE_C8,CRC_EMTAB8107_data_DE_C9, file = "CRC_EMTAB8107_data_DE_C6_8_9.RData")
load("CRC_EMTAB8107_data_DE_C6_8_9.RData")
save(CRC_GSE166555_data_DE_C5,CRC_GSE166555_data_DE_C11,CRC_GSE166555_data_DE_C13,CRC_GSE166555_data_DE_C14,CRC_GSE166555_data_DE_C18,CRC_GSE166555_data_DE_C29, file = "CRC_GSE166555_data_DE_C5_11_13_14_18_29.RData")
load("CRC_GSE166555_data_DE_C5_11_13_14_18_29.RData")
